finalized README

+++ db/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE departments (
 CREATE TABLE roles (
   id SERIAL PRIMARY KEY,
   title VARCHAR(30) UNIQUE NOT NULL,
-  salary DECIMAL NOT NULL,
+  salary INTEGER NOT NULL,
   department_id INTEGER NOT NULL,
   FOREIGN KEY (department_id) REFERENCES departments(id)
 );
